Added meter support

+++ src/main/java/se/anviken/owmanager/model/Meter.java
@@ -1,33 +1,41 @@
 package se.anviken.owmanager.model;
 
 import java.io.Serializable;
-import javax.persistence.*;
 import java.util.Date;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 
 /**
  * The persistent class for the meters database table.
  * 
  */
 @Entity
-@Table(name="meters")
-@NamedQuery(name="Meter.findAll", query="SELECT m FROM Meter m")
+@Table(name = "meters")
+@NamedQuery(name = "Meter.findAll", query = "SELECT m FROM Meter m")
 public class Meter implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@Column(name="meter_id")
+	@Column(name = "meter_id")
 	private int meterId;
 
 	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name="meter_timestamp")
+	@Column(name = "meter_timestamp")
 	private Date meterTimestamp;
 
-	private double value;
+	private float value;
 
-	//bi-directional many-to-one association to MeterType
+	// bi-directional many-to-one association to MeterType
 	@ManyToOne
-	@JoinColumn(name="meter_type_id")
+	@JoinColumn(name = "meter_type_id")
 	private MeterType meterType;
 
 	public Meter() {
@@ -49,11 +57,11 @@ public class Meter implements Serializable {
 		this.meterTimestamp = meterTimestamp;
 	}
 
-	public double getValue() {
+	public float getValue() {
 		return this.value;
 	}
 
-	public void setValue(double value) {
+	public void setValue(float value) {
 		this.value = value;
 	}
